fix: move access code, fix import export
