--- src/accessCode/index.ts
@@ -1,5 +1,0 @@
-import AccessCodeProvider from './accessCodeProvider'
-
-export const AccessCode = {
-  AccessCodeProvider,
-}
