+++ src/components/accessCode/index.ts
@@ -1,0 +1,1 @@
+export { default as AccessCodeProvider } from './accessCodeProvider'
