+++ src/components/index.ts
@@ -1,3 +1,4 @@
+export * from './accessCode'
 export * from './button'
 export * from './font'
 export * from './bottomsheet'
